Move the height to a double for better interpolatory goodness

+++ landscape.h
@@ -4,7 +4,7 @@
 #include <sys/queue.h>
 
 /* We need to keep track of active vs inactive points */
-typedef struct { int x; int y; int z; } point;
+typedef struct { int x; int y; double z; } point;
 struct pq_entry { point p; TAILQ_ENTRY(pq_entry) entries; };
 struct pq_entry *np, *np_temp;
